+++ yandex/resource_yandex_vpc_security_group_test.go
@@ -270,6 +270,7 @@ resource "yandex_vpc_security_group" "sg1" {
     description    = "rule1 description"
     protocol       = "ICMP"
     v4_cidr_blocks = ["10.0.1.0/24", "10.0.2.0/24"]
+    port = -1
   }
 
   ingress {